"重构服务器模块结构，分离下载器、日志和WebSocket服务实现"
