+++ python_app/__init__.py
@@ -1,6 +1,7 @@
 """Python desktop/WebSocket downloader package."""
 
 from .config import ServerConfig
-from .server import WebSocketDownloadServer, run_server_forever
+from .logging_utils import configure_logging
+from .websocket_server import WebSocketDownloadServer, run_server_forever
 
-__all__ = ["ServerConfig", "WebSocketDownloadServer", "run_server_forever"]
+__all__ = ["ServerConfig", "WebSocketDownloadServer", "run_server_forever", "configure_logging"]
