+++ python_app/locales/__init__.py
@@ -1,0 +1,3 @@
+"""Local JSON translation resources for the desktop client."""
+
+__all__ = []
